requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
google-genai>=0.2.0
google-auth>=2.0.0
google-cloud-firestore>=2.0.0
google-cloud-storage>=2.0.0
google-cloud-secret-manager>=2.0.0
google-cloud-texttospeech>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
fastapi
uvicorn[standard]
requests>=2.31.0
pytest>=7.4.0
yfinance>=0.2.0
Pillow>=10.0.0


